% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntDigitsCounter.R
\name{DigitCounter}
\alias{DigitCounter}
\title{Count numbers of individual integer digits in a vector}
\usage{
DigitCounter(x, na.rm = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{x}{Numeric vector}

\item{na.rm}{A logical input indicating whether missing values should be removed before the computation proceeds.}

\item{verbose}{A logical input indicating whether keeps the user up-to-date with messages that print to screen as the function is running.}

\item{...}{Other arguments passed on to methods}
}
\value{
A vector containing the numbers of integer digits of each number in input
}
\description{
Count numbers of individual integer digits in a vector
}
\examples{
DigitCounter_single(43458.99, verbose=TRUE)
DigitCounter(c(0.123, 1, -245, 4308.99), na.rm = FALSE)
}
