% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntDigitsCounter.R
\name{IntDigitsCounter}
\alias{IntDigitsCounter}
\title{Count Numbers of Individual Integer Digits in a Vector}
\usage{
IntDigitsCounter(x, na.rm = TRUE, verbose = FALSE, ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{na.rm}{A logical input indicating whether missing values should be removed before the computation proceeds.}

\item{verbose}{A logical input indicating whether keeps the user up-to-date with messages that print to screen as the function is running.}

\item{...}{Other arguments passed on to methods.}
}
\value{
A vector containing the numbers of integer digits of each number in input.
}
\description{
Count Numbers of Individual Integer Digits in a Vector
}
\examples{
IntDigitsCounter(c(0.123, 1, -245, 4308.99), na.rm = FALSE)
}
